+++ web/src/components/ZDivider/ZDivider.vue
@@ -1,8 +1,7 @@
 <template lang="pug" functional>
   .z-divider(
     :ref="data.ref"
-    :class="[data.class, data.staticClass, props.light ? 'z-divider--light' : 'z-divider--dark']"
-    :style="[data.style, data.staticStyle]"
+    :class="[data.class, data.staticClass, $options.methods.computeClasses(props)]"
     :style="[data.style, data.staticStyle, $options.methods.computeStyle(props)]"
     v-on="listeners"
   )
@@ -15,14 +14,31 @@
   width 100%
   height 1px
   margin size(1) 0
+  overflow hidden
+  transition height $transform-duration $ease
 
   &--dark
-    background-color: $colors.secondary.base
-    opacity 0.25
+    background-color: alpha($colors.secondary.base, 25%)
 
   &--light
-    background-color: $colors.white.base
-    opacity 0.5
+    background-color: alpha($colors.white.base, 50%)
+
+  &--loading
+    &:before
+      content ''
+      display block
+      width 100%
+      height 100%
+      background-color: $colors.primary.base
+      transform translate(50px)
+      animation z-divider-loading-animation 2.6s $ease infinite
+
+@keyframes z-divider-loading-animation
+  0%
+    transform translate(-100%)
+
+  100%
+    transform translate(100%)
 </style>
 
 <script>
@@ -35,12 +51,22 @@ export default {
       type: Boolean,
       default: false,
     },
+    loading: {
+      type: Boolean,
+      default: false,
+    },
     height: {
       type: [String, Number],
       default: '1px',
     },
   },
   methods: {
+    computeClasses(props) {
+      const classes = [];
+      classes.push(props.light ? 'z-divider--light' : 'z-divider--dark');
+      if (props.loading) classes.push('z-divider--loading');
+      return classes;
+    },
     computeStyle(props) {
       return sizeStyle(props.height, { width: false });
     },